+++ nodeserver/countdown.js
@@ -1,5 +1,5 @@
 // countdown.js
-const logger = require('../shared/build/logger');
+const logger = require('../shared/build/logger').default;
 const { connectToRedis } = require('../shared/build/redisClient');
 const MqttPublisher = require('./mqtt/mqttPublisher');