+++ nodeserver/routes/getSecretsRoute.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const vaultClient = require('../../shared/build/vaultClient'); // Ensure the path is correct
-const logger = require('../../shared/build/logger');
+const logger = require('../../shared/build/logger').default;
 
 router.get('/', async (req, res) => {
     try {